bpf: add -DIPV4_FRAGMENTS to BPF Makefile
Add the options for IPv4 frament tracking support to the Makefile, so
that BPF programs are built (and loaded, depending on the case) with the
added complexity caused by fragment tracking support.

Do not add it to MAX_LB_OPTIONS or MAX_LXC_OPTIONS, as the tests
involving those are currently run by CI on Linux 4.9 only (with no LRU
map support - see #10798).

Signed-off-by: Quentin Monnet <quentin@isovalent.com>